custom config module called inside the check logic
